Route normalization
Each background route is normalized to 11 points, which in some cases
increases the number of points in the route but in many cases decreases
them. This is done to better resolve cartogram distortion, which steps
through each point in a route and linearly distorts it based on a ramp
between the costs of each end of the route. This was causing some
graphical representation issues with routes that were particularly
complicated on one side or another of the route.